Move cat command into own file
